# apps/web/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Strategy Terminal",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/market/quote": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Query Market Quote",
        "operationId": "queryMarketQuote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/quote/refresh": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Refresh Market Quote",
        "operationId": "refreshMarketQuote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteRefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/bars/query": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Query Market Bars",
        "operationId": "queryMarketBars",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketBarsView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analysis/snapshots": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Analysis Snapshot",
        "operationId": "createAnalysisSnapshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Request or analysis input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Data or service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analysis/research": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Research Analysis",
        "operationId": "createResearchAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request or analysis input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Data or service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analysis/realtime-preview": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Realtime Analysis Preview",
        "operationId": "createRealtimeAnalysisPreview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealtimeAnalysisPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeAnalysisPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Preview input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Preview data unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analysis/structures": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Technical Structure Analysis",
        "operationId": "createTechnicalStructureAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicalStructureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalStructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request or analysis input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Structure service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/instruments/search": {
      "get": {
        "tags": [
          "instruments"
        ],
        "summary": "Search Instruments",
        "operationId": "searchInstruments",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Adjustment": {
        "type": "string",
        "enum": [
          "forward",
          "backward",
          "none"
        ],
        "title": "Adjustment"
      },
      "AnalysisRequest": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/DataQuery"
          },
          "benchmark_query": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataQuery"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "has_position": {
            "type": "boolean",
            "title": "Has Position",
            "default": false
          },
          "provisional_quote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvisionalQuoteInput-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisSnapshot": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "timeframe": {
            "$ref": "#/components/schemas/Timeframe"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "bar_status": {
            "$ref": "#/components/schemas/SignalStatus"
          },
          "engine_version": {
            "type": "string",
            "minLength": 1,
            "title": "Engine Version"
          },
          "config_version": {
            "type": "string",
            "minLength": 1,
            "title": "Config Version"
          },
          "data_lineage": {
            "items": {
              "$ref": "#/components/schemas/DataLineage"
            },
            "type": "array",
            "title": "Data Lineage"
          },
          "market_regime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketRegimeContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "sector_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectorOpportunity"
              },
              {
                "type": "null"
              }
            ]
          },
          "pair_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PairCandidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "component_scores": {
            "items": {
              "$ref": "#/components/schemas/ScoreComponent"
            },
            "type": "array",
            "title": "Component Scores"
          },
          "evidence_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Evidence Score"
          },
          "action": {
            "$ref": "#/components/schemas/DecisionAction"
          },
          "key_levels": {
            "items": {
              "$ref": "#/components/schemas/KeyLevel"
            },
            "type": "array",
            "title": "Key Levels"
          },
          "invalidation": {
            "items": {
              "$ref": "#/components/schemas/Invalidation"
            },
            "type": "array",
            "title": "Invalidation"
          },
          "explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument",
          "timeframe",
          "as_of",
          "bar_status",
          "engine_version",
          "config_version",
          "data_lineage",
          "component_scores",
          "evidence_score",
          "action"
        ],
        "title": "AnalysisSnapshot"
      },
...
      "DataQuery": {
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "timeframe": {
            "$ref": "#/components/schemas/Timeframe"
          },
          "purpose": {
            "$ref": "#/components/schemas/DataPurpose",
            "default": "research"
          },
          "adjustment": {
            "$ref": "#/components/schemas/Adjustment",
            "default": "forward"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument",
          "timeframe"
        ],
        "title": "DataQuery"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstrumentId": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "asset_type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "CNY"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "market",
          "asset_type"
        ],
        "title": "InstrumentId"
      },
...
      "InstrumentSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstrumentSearchItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "items"
        ],
        "title": "InstrumentSearchResponse"
      },
...
      "MarketBarsView": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "market-bars-view/v1",
            "title": "Schema Version"
          },
          "data_fingerprint": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Data Fingerprint"
          },
          "data": {
            "$ref": "#/components/schemas/DataEnvelope"
          },
          "indicator_profile": {
            "$ref": "#/components/schemas/CoreIndicatorProfileView"
          },
          "indicators": {
            "$ref": "#/components/schemas/CoreIndicatorsView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "data_fingerprint",
          "data",
          "indicator_profile",
          "indicators"
        ],
        "title": "MarketBarsView"
      },
...
      "QuoteQuery": {
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "purpose": {
            "$ref": "#/components/schemas/DataPurpose",
            "default": "research"
          },
          "temporal_mode": {
            "$ref": "#/components/schemas/QuoteTemporalMode",
            "default": "exact"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "max_live_age_seconds": {
            "type": "number",
            "maximum": 900.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Live Age Seconds",
            "default": 90.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument"
        ],
        "title": "QuoteQuery",
        "description": "Explicit temporal request for a native provider quote snapshot."
      },
      "QuoteRefreshRequest": {
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument"
        ],
        "title": "QuoteRefreshRequest"
      },
      "QuoteRefreshResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "quote-refresh/v1",
            "title": "Schema Version",
            "default": "quote-refresh/v1"
          },
          "quote": {
            "$ref": "#/components/schemas/QuoteEnvelope"
          },
          "provider_unchanged": {
            "type": "boolean",
            "title": "Provider Unchanged"
          },
          "cooldown_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Cooldown Seconds"
          },
          "origin": {
            "type": "string",
            "const": "manual_refresh",
            "title": "Origin",
            "default": "manual_refresh"
          },
          "cache_bypassed": {
            "type": "boolean",
            "const": true,
            "title": "Cache Bypassed",
            "default": true
          },
          "alert_event_created": {
            "type": "boolean",
            "const": false,
            "title": "Alert Event Created",
            "default": false
          },
          "external_notification_sent": {
            "type": "boolean",
            "const": false,
            "title": "External Notification Sent",
            "default": false
          },
          "trade_created": {
            "type": "boolean",
            "const": false,
            "title": "Trade Created",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quote",
          "provider_unchanged",
          "cooldown_seconds"
        ],
        "title": "QuoteRefreshResponse"
      },
...
      "QuoteViewResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "quote-view/v2",
            "title": "Schema Version"
          },
          "quote": {
            "$ref": "#/components/schemas/QuoteEnvelope"
          },
          "formal_use_eligible": {
            "type": "boolean",
            "title": "Formal Use Eligible"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "quote",
          "formal_use_eligible"
        ],
        "title": "QuoteViewResponse"
      },
...
      "RealtimeAnalysisPreviewRequest": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/DataQuery"
          },
          "benchmark_query": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataQuery"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "has_position": {
            "type": "boolean",
            "title": "Has Position",
            "default": false
          },
          "provisional_quote": {
            "$ref": "#/components/schemas/ProvisionalQuoteInput-Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "provisional_quote"
        ],
        "title": "RealtimeAnalysisPreviewRequest",
        "description": "A full provisional analysis request bound to one latest native quote."
      },
...
      "ResearchAnalysisResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "research-view/v1",
            "title": "Schema Version"
          },
          "data_ref": {
            "$ref": "#/components/schemas/ResearchDataReferenceView"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AnalysisSnapshot"
          },
          "market_view": {
            "$ref": "#/components/schemas/MarketBarsView"
          },
          "technical_summary": {
            "$ref": "#/components/schemas/TechnicalSummaryView"
          },
          "section_status": {
            "$ref": "#/components/schemas/ResearchSectionStatusView"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "realtime_signal_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RealtimeSignalPreviewView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "data_ref",
          "snapshot",
          "market_view",
          "technical_summary",
          "section_status",
          "warnings"
        ],
        "title": "ResearchAnalysisResponse"
      },
...
    }
  }
}